Do a better job of localizing book names
